feat(GameResult): 당첨 통계에서 통해 게임의 수익률을 확인할 수 있다

+++ src/main/kotlin/lotto/LottoShop.kt
@@ -8,7 +8,7 @@ object LottoShop {
     fun sellLottos(cash: Money): Lottos {
         val amountOfLotto = cash.amount / LOTTO_PRICE.amount
         val generatedLottos = mutableListOf<Lotto>()
-        repeat(amountOfLotto) {
+        repeat(amountOfLotto.toInt()) {
             generatedLottos.add(generateLotto())
         }
